Fix 0x90 byte literal again (need .toByte() for extended value test)

+++ app/src/test/java/com/example/ava/esphome/infrared/IrCodecTest.kt
@@ -69,7 +69,7 @@ class IrCodecTest {
             38,                 // type
             2,                  // repeat -> repeat+1 = 3
             9, 0,               // LE length = 9 (3 pulses + 2 extended bytes + 2 end marker)
-            0x00, 0x90, 0x01,   // extended value 0x0190 = 400
+            0x00, 0x90.toByte(), 0x01,   // extended value 0x0190 = 400
             10,                 // single byte
             0x00, 0x07, 0x00,   // extended value 7 (LE 0x0007)
             0x0D, 0x05,         // end marker
